requirements.txt
langchain==0.0.267
langchain-community
openai==0.27.8
chromadb==0.4.13
python-dotenv==1.0.0
streamlit==1.26.0
pandas==2.0.3
python-docx==0.8.11
pyPDF2==3.0.1
faiss-cpu==1.7.4
tiktoken==0.4.0
matplotlib==3.7.2
plotly==5.16.1
fpdf==1.7.2